requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.1
django-bootstrap4==2.2.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-simple-history==2.11.0
django-tables2==2.3.1
mysqlclient==2.0.1
Pillow==7.2.0
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
